Handle bmson sound offsets correctly

+++ src/gameplay_logic/BmsGameReferee.cpp
@@ -8,11 +8,11 @@
 gameplay_logic::BmsGameReferee::BmsGameReferee(
   std::array<std::vector<charts::BmsNotesData::Note>,
              charts::BmsNotesData::columnNumber> notes,
-  const std::vector<std::pair<charts::BmsNotesData::Time, uint16_t>>& bgmNotes,
+  const std::vector<std::pair<charts::BmsNotesData::Time, uint64_t>>& bgmNotes,
   std::vector<std::pair<charts::BmsNotesData::Time, double>> bpmChanges,
   std::shared_ptr<sounds::Sound> mineHitSound,
   BmsLiveScore* score,
-  std::unordered_map<uint16_t, std::shared_ptr<sounds::Sound>> sounds,
+  std::unordered_map<uint64_t, std::shared_ptr<sounds::Sound>> sounds,
   rules::HitRules hitRules)
   : bpmChanges(std::move(bpmChanges))
   , sounds(std::move(sounds))
